fix power limit usage
Some laptops and/or older graphics aren't compatible with querying the power management limit.  In such circumstances, it'll attempt to gather the upper limit using a different parameter within the pynvml library, and if that fails, it'll explicitly set "none" so as to not prevent all other metrics from being gathered and displayed.