fix ChatTTS - v6.0.3
Temporarily fixes the issue of correctly specifying the path to the local model when using the "local" parameter instead of "huggingface" - see comments in the script itself.  This was necessary because the ```ChatTTS``` library doesn't expose the ```cache_dir``` parameter that's part of the```huggingface_hub``` library's ```snapshot_download``` method.  This workaround sort of solves the issue except that the folder within the TTS folder isn't named like other times I've used "cache_dir"...but in order to make everything uniform, currently, it'd require modifying the ```ChatTTS``` source code.