use sqlite3
Create an sqlite3 database when creating the vector database.  The new database will contain relevant page content, file name, file path, and has value for all files put into the vector database.  This is necessary to populate the Manage Databases tab because the prior method of parsing JSON files was horribly slow when the vector database contains, for example, 120,000 documents...