fix 2 errors
Previously, the script always used AV to pre-process an audio file and always attempted to delete the temporary .wav file that was created.  An error showed that the temporary .wav couldn't bet deleted under the new approach, which uses ffmpeg within whispers2t for pre-processing instead.  This fixed that.

It also removed a redundant print statement that's already in gui_tabs_tools_transcribe.mpy.